program/requirements.txt
prometheus_client==0.21.1 --pre

# recupérer les info du pc
psutil==7.0.0 --pre

#pour les gpio
lgpio==0.2.2.0 --pre
board==1.0 --pre
adafruit-circuitpython-ina219==3.4.26 --pre
gpiod==2.3.0 --pre

# pour la communication avec le servers
requests==2.32.3 --pre
websocket-client==1.8.0 --pre
python-socketio==5.12.1 --pre

#seulement pour la raspberry pi 4
#RPi.GPIO==0.7.1 --pre